fix(hmr): run HMR handler sequentially
